bench: repeat using an extern crate bit_vec
This splits the mod bench into its own file, and then adds a separate
benchmark program to also run the same tests.  This way we can see the
difference between running within the bit_vec crate vs. external.
